+++ MangAdventure/__init__.py
@@ -1,3 +1,3 @@
 """A simple manga hosting CMS written in Django."""
 
-__version__ = '0.9.5'
+__version__ = '0.9.6'