+++ scripts/lint.sh
@@ -2,9 +2,9 @@
 
 declare -i e=0
 
-flake8 || ((++e))
-isort -q -c --df . || ((++e))
-mypy --no-error-summary . || ((++e))
-bandit -c pyproject.toml -r . || ((++e))
+flake8 || ((e+=1))
+isort -q -c --df . || ((e+=2))
+mypy --no-error-summary . || ((e+=4))
+bandit -q -c pyproject.toml -r . || ((e+=8))
 
 exit $e